ARCH-3: Moved all the subviews to their superviews namespace by declaring it in superview's namespace
