--- WineUp/WineUp/Views/CatalogFiltersBar/CatalogFiltersBarItemButton.swift
@@ -1,41 +1,0 @@
-//
-//  CatalogFiltersBarItemButton.swift
-//  WineUp
-//
-//  Created by Александр Пахомов on 27.09.2020.
-//
-
-import SwiftUI
-
-// MARK: - View
-
-/// `CatalogFiltersButtonItemView` wrapper with `onTap` callback
-struct CatalogFiltersBarItemButton: View {
-
-    let item: CatalogFiltersBarView.Item
-    let onTap: (() -> Void)?
-
-    var body: some View {
-        Button(action: didTap, label: {
-            CatalogFiltersBarItemView(item: item).padding()
-        })
-    }
-}
-
-// MARK: - Helpers
-
-private extension CatalogFiltersBarItemButton {
-    func didTap() {
-        onTap?()
-    }
-}
-
-// MARK: - Preview
-
-#if DEBUG
-struct CatalogFiltersBarItemButton_Previews: PreviewProvider {
-    static var previews: some View {
-        CatalogFiltersBarItemButton(item: CatalogFiltersBarView.Item.mockedData[0], onTap: nil)
-    }
-}
-#endif
